+++ src/wheel_inspect/errors.py
@@ -3,17 +3,6 @@ class WheelValidationError(Exception):
     pass
 
 
-class InvalidFilenameError(WheelValidationError, ValueError):
-    """ Raised when an invalid wheel filename is encountered """
-
-    def __init__(self, filename):
-        #: The invalid filename
-        self.filename = filename
-
-    def __str__(self):
-        return 'Invalid wheel filename: ' + repr(self.filename)
-
-
 class RecordValidationError(WheelValidationError):
     """
     Superclass for all validation errors raised due to a wheel's :file:`RECORD`